+++ src/main/java/gregtech/GT6_Main.java
@@ -51,6 +51,7 @@ import gregtech.blocks.fluids.BlockOcean;
 import gregtech.blocks.fluids.BlockRiver;
 import gregtech.blocks.fluids.BlockSwamp;
 import gregtech.compat.*;
+import gregtech.compat.waila.CompatWaila;
 import gregtech.entities.projectiles.EntityArrow_Material;
 import gregtech.entities.projectiles.EntityArrow_Potion;
 import gregtech.items.tools.early.GT_Tool_Scoop;
@@ -263,6 +264,7 @@ public class GT6_Main extends Abstract_Mod {
 		new Compat_Recipes_ActuallyAdditions    (MD.AA            , this);
 		new Compat_Recipes_ExtraUtilities       (MD.ExU           , this);
 		new Compat_Recipes_WRCBE                (MD.WR_CBE_C      , this);
+		new CompatWaila                         (MD.WAILA         , this);
 		
 		new CompatMods(MD.GT, this) {@Override public void onPostLoad(FMLPostInitializationEvent aInitEvent) {
 			ArrayListNoNulls<Runnable> tList = new ArrayListNoNulls<>(F,