+++ src/main/java/gregtech/compat/waila/GTWailaUtils.java
@@ -1,0 +1,7 @@
+package gregtech.compat.waila;
+
+import java.text.DecimalFormat;
+
+public class GTWailaUtils {
+    public static DecimalFormat DF1 = new DecimalFormat("#.###"); // 123.4567 -> 123.456, 123.0000 -> 123
+}